pyproject.toml
[project]
name = "williott"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.122.0",
    "google-cloud-texttospeech>=2.33.0",
    "hypermedia>=5.3.4",
    "psycopg[binary]>=3.2.13",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "sqlmodel>=0.0.27",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "ruff>=0.14.7",
]
